removed submission-service refactored submission.js
