+++ test/graphql-queries/batch-job.test.js
@@ -8,10 +8,10 @@ const {EmailService} = require("../../services/email");
 const {NotifyUser} = require("../../services/notify-user");
 const {User} = require("../../crdc-datahub-database-drivers/services/user");
 const {ApprovedStudiesService} = require("../../services/approved-studies");
-const {SubmissionService} = require("../../crdc-datahub-database-drivers/services/submission-service");
 const {S3Service} = require("../../crdc-datahub-database-drivers/services/s3-service");
 const {BatchService} = require("../../services/batch-service");
 const ERROR = require("../../constants/error-constants");
+const {Submission} = require("../../services/submission");
 jest.mock("../../crdc-datahub-database-drivers/mongodb-collection");
 jest.mock("../../crdc-datahub-database-drivers/mongo-queries.js");
 jest.mock("../../crdc-datahub-database-drivers/services/user");
@@ -26,7 +26,7 @@ const notificationsService = new NotifyUser(emailService);
 const userService = new User(userCollection);
 
 const submissionCollection = new MongoDBCollection();
-const submissionService = new SubmissionService(submissionCollection);
+const submissionService = new Submission(submissionCollection);
 const s3Service = new S3Service();
 const batchCollection = new MongoDBCollection();
 const batchService = new BatchService(s3Service, batchCollection, config.submission_aws_bucket_name);
